Improvements to link trail handling, and two tweaks to the whitelist. 182 tests now passing.
Link trails depend on language-dependent positive character classes in the PHP
parser. These classes all seem to disallow punctuation implicitly and list
differing plain text characters instead, so it might be possible to get away
with identifying a common class of non-trail punctuation instead. This would
help to keep the tokenizer independent of configurations, which is very
desirable for caching and simplified external parsing.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions/VisualEditor@107633 dd0e9695-b195-4be7-bd10-2dea1a65a6b6